pyproject.toml
[project]
name = "ai-wrapper-api"
version = "0.1.0"
description = "FastAPI backend for the AI Wrapper virtual user research UI."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "openai>=1.30.0",
  "pydantic-settings>=2.3.0",
  "tenacity>=8.2.3",
  "numpy>=1.26.4"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0"
]

[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
